[lex] Provide unicode name for all control characters
This commit does not touch the new-line character as paper
P2348.  It resricts itself to consistent use of the unicode
character name for space, horizontal tab, and vertical tab.
Compared to PR #7359 it deliberately does not touch the
grammar that would necessitate a review by core review.
The intent is to rebase that PR if this one lands.