[dcl.spec.auto.general]-p2: wording on param. type
Avoid incorrect sentence parsing option on parameter type.
Text may be ambiguous in the sense that it may be interpreted both as the type of any of three things (incorrect) or the type of a parameter of any of three things (correct).
Despite the former interpretation does not make sense as a lambda expression has never an explicit type, it would still be useful to avoid ambiguity in sentence parsing.

Signed-off-by: Mauro Russo <mrusso@luxoft.com>