Use template-parameter and template parameter more consistently
Try to use template-parameter only when we refer to a
grammar construct, and to 'template parameter' everywhere else.

Adopt the same logic to template-argument/template argument.

This change might not be  exhaustive.

The aim is to editorially adopt some of the wording changes
made in P2841R5 to ease its review in core.