[cpp] Correctly identify preprocessor tokens vs, tokens
Most references to tokens in [cpp] correctly refer to preprocessing
tokens that a materially different thng to a regular token.  This
commit addresses the last few places referring to a plain token
where they should instead be referring to a preprocessing token.
The remining usage of plain token where evaludating the predicate
to  directives is intentional, and trying to change to a
preproccessing token would be an actual error.