[macros] Add LaTeX macros to index library macros
The immediate idea is to support using the new macros directly
in header synopses when defining each library macro.  This will
ensure that no macros are accidentally not indexed.

A follow-up plan is that this separation of library macros will
make it easier to create a separate index of macros, or apply
other macro-specific renderings, in the future.  To this end,
all indexed uses of a macro, not just those in header files,
should be replaced by use of these new macros.  Similarly,
these LaTeX macros can be used in-place in regular text to
index cross-references where standard library macros are used
throughout the standard.