[temp.mem.enum] Fix consistency with [temp.inst]
The example is inconsistent with [temp.inst] p3, which states the following:
"The implicit instantiation of a class template specialization causes
 - the implicit instantiation of the declarations, but not of the definitions,
   of [...] scoped member enumerations [...]; and
 - the implicit instantiation of the definitions of [...]
   unscoped member enumerations [...]"

Hence, the example can be fixed by either (1) omitting the instantiation or
(2) by replacing the unscoped with a scoped enumeration.
Since the implicit instantiation does not contribute to the point of
[temp.mem.enum], the inconsistency can be solved most easily with alternative (1).

Note: The example was originally taken from DR1206.