[lex.phases] Update implementation defined text
Since C++23 we no longer have physical source files, but rather
input files.  Update the two implementation-defined references
to the mapping from input file to translation character set
using the same phrasing so that they provide the same entry
in the index of implementation-defined behavior, just as they
did in C++20, before getting out of sync when the terminology
changed.