[basic.pre] Defragment specification of names and entities
The current contents of [basic.pre] jump between specifying
different things.  This PR moves all the specification of
names to the front, followed by the specification of entities.

There are two main benefits: (1) the specification for when
two names are the same is a list of 4 rules that correspond
to the 4 things than can form a name --- the connection is
much clearer when the paragraphs are adjacent and the list
is sorted to the same order; (2) in this form, even though
all the words are the same, the reordering and merging of
paragraphs a fit on a single page.  The very last paragraph
was forced over a page-break in the original layout.