Case where resumption does not resume as if the function is invoked
Regarding `breakpoint`:
The specific claim that the debugger resumes execution of the program as
if the function was not invoked is confusing considering that the
debugger may effect side-effects or cause execution to resume from a
different evaluation.

Instead, the idea is that `breakpoint` is not responsible for causing
the translation process to make special accomodations for resumption of
execution other than in cases where the debugger was strictly used for
observation only.