[debugging.utility] Clarify wording in notes
The previous wording in the notes in `breakpoint` and `is_debugger_present`
read as statements of fact about the implementation-defined behaviour.
The statements are actually ones of intent.

The specific claim in `breakpoint` that the debugger resumes execution of the program
as if the function was not invoked is confusing considering that the debugger may effect
side-effects or cause execution to resume from a different evaluation.

Instead, the idea is that `breakpoint` is not responsible for causing the translation process
to make special accomodations for resumption of execution other than in cases
where the debugger was strictly used for observation only.

In `is_debugger_present`, the functionality ascribed to POSIX by the wording
("ptrace") is not present in POSIX. Update to reference the LSB and to use
the corresponding terminology ("tracing process").

The wording implies a preference to return `true` in case it is unknown
whether a debugger is present. Add a critical "only" to fix that.