[std, macros] Use "environ" for  grammar extraction.
This replaces the use of the "extract" package with the simpler
"environ" package, where the grammar contents are extracted from the
"collected body" \BODY macro via the TeX `\meaning` command. This does
not preserve whitespace as the "extact" approach did, but if we just
spell out whitespace (instead of relying on \obeyspaces), we get the
same results.