[expected.object.cons] Reorder arguemnts of is_same_v for consistency
During the discussion of LWG 4222 it was noticed that [expected.object.cons] bullet (23.2) uses an atypical order of the is_same_v arguments, especially since this issue adds another bullet. To enforce consistency, the `remove_cvref_t<U>` argument is no also used as first (instead of second) argument in that bullet.