P3037R6 constexpr std::shared_ptr and friends
Fixes #7941
Fixes cplusplus/papers#1713

Did not apply `constexpr` to deleted functions. The application
in the paper was inconsistent, and I deemed the active uses as
errors.

Made minor whitespace and line-breaking adjustments to avoid
line-wrapping and overfull hboxes.

Fixed odd use of return type on the same line as the template
head for an `allocate_shared` overload, where it belonged on the
same line as the function name, before applying the `constexpr`
keyword.