[lex.separate] Redistribute second comment
The second comment in [lex.separate] is both too specific, and not specific
enough.  Move the last sentence about use of separately translated TUs to
the end of the first note, where it seems most appropriate.  Move the rest
to just after the definition of translation unit in phase 7 of translation.