[lex.phases] Move dropping whitespace to end of phase 4
Phase 7 deals only with tokens, not whitespace.  For clarity
move the dropping of whitespace to the end of phase 4, after
preprocessing directives are deleted.  Newline characters are
significant in the grammar of preprocessing directives to
whitespace characters are clearly signicant until this point.
Conversely, it is good to recognise that string literal
preprocessing tokens are adjacted regardless of whitespace
in phases 5 and 6.