[dcl.inline] inline specifier is for ODR
The key use of the inline specifier since C++11, if not before,
has been to allow multiple declarations to satisfy the ODR rather
than to provide a hint that compilers routinely ignore.

This change moves but does not change wording, in order to move
the comment making the connection with the ODR more prominent
than the normative wording suggestings core transformation.

It might be desirable to demote the normative coding hint to
a note, but that goes beyond the remit of a simple editorial
chsnge.