[macros] Restore sentence spacing after `\tcode`
TeX treats a word ending with an uppercase letter as an abbreviation.
This makes the space after `\tcode{T}.` result in regular interword
spacing. To restore proper (wider) sentence spacing, this commit
appends `\@` to the `\tcode` command. The insertion of `\@` happens
only during normal typesetting; it is suppressed when writing the .idx
file and when generating bookmarks.