[lib.intro, iterator.concept.inc, concept.regularinvocable] Avoid "annotation"; hyphenate "freestanding-deleted"
Now that many sections use "annotation" to mean a specific core-language
construct, we should avoid using "annotation" to mean anything but that construct,
i.e. stop using it to mean "code comment." (There will be an NB comment to this effect.)

Drive-by hyphenate "freestanding-deleted function". Not all freestanding-deleted
functions are freestanding functions, and not all freestanding-deleted functions
are deleted functions; hyphenating helps the reader understand that this is a
term of art that cannot meaningfully be broken down into its component notions.