[meta.unary.prop] [meta.rel] P/R for LWG3967
This intends to eliminate all instances of "is known not to throw
any exception" from the library clauses. "Is known" is too vague
a condition (known to whom?); we ought to define it explicitly
in terms of potentially-throwing expressions, i.e., we expect these
traits to exactly match the result of the `noexcept` operator.