[cpp.line] Simplify specification of line directives
This change set addresses a variety of concerns to produce
a simpler and clearer specification for line directives.

The first issue to address is that the specification refers
to a _string-literal_ that is not present in the grammar.
Howeer, the constraints on that string literal are sufficient
to replace that quoted _s-char-sequence_ in the grammar
production with _string-literal_.

The line directive itself is given its own grammar production
that formally states what used to be described in text
restrictions and interprettions of the generic sequence of
preprocessing tokens.

Refactoring the grammar as its own production provides a
simpler way for incoming proposals to refer to line
directives, such as https://wg21.link/P3868.