[cpp.line] Address string-literal vs s-char-sequence
The text of this specification places a restriction on a string-literal
that is not present in the grammar, although the grammar is an exact
match for a string-literal with that restriction.

The restricted term is also used when specifying the effects of the
line directive with a string, so prefer to move string-literal into
the grammar rather than strike that paragraph and adjust the effects.

Added a cross-reference to where character string literal is defined
as it may not be obvious to all readers that this is a term of power,
even though it is defined as a term and has an index entry.