[lex.string] Adopt character-string grammer to simplify string literals
There are several places in the grammar where we want to refer to a
plain character string literal without a prefix and that is not raw.
Introduce _character-string_ into the grammar to replace a quoted
_s-char-sequence_in a number of places, and move the `R` of a raw
string literal into the _raw-string_ grammar so that _raw-string_
and _character-string_ are used consistently where they appear.