[lex] Complete the use of unicode code points
Completes the task of applying Unicode code point markup to
refer to single characters in normative text.

Replace all remaining uses of backslash as glyph (\) or text
with the corresponding `unicode{005c}{reverse solidus}` markup.

Replace a couple of text reference to single and double quotes
with their corresponding Unicode markup, and similarly for the
glyphs ` and ".

This should compete the last Unicode markup in [lex].