[iterator.range] Turn list of headers into a table
It is tricky to properly line-break the list of additional headers
that contain the range functions, and that list is growing quite
large to be displayed as a list in a single sentence.  Prefer to
render the list as a table instead.

Retained the cross-references to the header synopses from the
original text, but they look odd in the table format.