[lex.pptoken] Turn non-normative text into a note
The references to creating keyword placeholders for module
syntax are non-normative, simply pointing to the normative
text in [cpp].  Therefore, extend the note at the end of
the paragraph to cover the whole paragraph.  Then remove
the line number to attach the note to the preceding
paragraph on on making tokens out of preprocessing tokens.