[lex.phases] Make note on the notion of file more prominent
The notion that when the standard refers to files does not
necessarily imply a file in a traditional filing system is
more fundamental than its late appearance in the middle of
phase 7 of translation.  Move that note right to the top of
[lex] where we first talk of storing the program text in
source files, where is will be understood even before the
phases of translation.