[container.requirements] Complete index of container operations
Many container operations are specified entirely by the container
requirements, so those container members should be indexed into
this specification. There is precedent for many operations, but
the current indexing is haphazard and incomplete. This change set
thoroughly reviews and correctly indexes all itemdecl entries in
the requirements clause, and omits indexing only for the abstract
notion of container node types.